Renaming files and classes for parser
