+++ Server/src/controllers/FileParserController.ts
@@ -1,12 +1,12 @@
 import { Request, Response } from 'express';
-import { DataExtractionService } from '../services/dataExtraction/DataExtractionService'
+import { DataParserService } from '../services/Parser/DataParserService'
 
 /**
- * The dataExtractionController is responsible for preparing a request and extracting key information
+ * The FileParserController is responsible for preparing a request and extracting key information
  * from the request. The controller also finds the extension of the input file as processing is done based on it
  */
 
-export class dataExtractionController {
+export class FileParserController {
   private filePathOfUpload: string;
   private fileName: string
   private fileExtension: string
@@ -22,17 +22,17 @@ export class dataExtractionController {
       response.status(400).json("Request Body is empty.");
     }
     else {
-      let requestResponse: any = this.callDataExtractionService(this.filePathOfUpload, this.fileExtension, response);
+      let requestResponse: any = this.callDataParserService(this.filePathOfUpload, this.fileExtension, response);
       return requestResponse
     }
   }
 
 
-  private async callDataExtractionService(filePath: string, extension: string, response: Response): Promise<Response> {
+  private async callDataParserService(filePath: string, extension: string, response: Response): Promise<Response> {
 
-    let dataService = new DataExtractionService(extension, filePath);
+    let dataService = new DataParserService(extension, filePath);
     try {
-      let extractDataResponse: any = await dataService.extractData();
+      let extractDataResponse: any = await dataService.parseData();
       return response.status(extractDataResponse.statusCode).json(extractDataResponse.message);
     } catch (error) {
       return response.status(error.status).json(error.message);
