+++ Server/src/services/Parser/DataParserService.ts
@@ -2,14 +2,13 @@ import { JsonFileExtractorFactory } from './JsonFileHandlerFactory';
 import { IResponse } from '../../genericInterfaces/ResponsesInterface';
 import { BadRequest } from '@tsed/exceptions';
 import { FileExtractorFactory } from './FileHandlerFactory';
-import parse from 'csv-parse';
 import { CSVFileExtractorFactory } from './CSVFileHandlerFactory';
 
 /**
  * The methods in this class are only responsible for processing uploaded files. Input will be parsed 
  * then stored into its appropriate table in the database. 
  */
-export class DataExtractionService {
+export class DataParserService {
 
     private dataType: string
     private filePath: string
@@ -20,7 +19,7 @@ export class DataExtractionService {
         this.filePath = filePath
     }
 
-    async extractData(): Promise<IResponse> {
+    async parseData(): Promise<IResponse> {
 
         let requestResponse: IResponse = {} as any
         switch (this.dataType) {
